+++ tests/testthat/test-as_tdm.R
@@ -1,0 +1,7 @@
+context("Checking as_tdm")
+
+test_that("as_tdm makes a TermDocumentMatrix",{
+
+    expect_true(methods::is(as_tdm(markers), "TermDocumentMatrix"))
+})
+